feat(harness): extract templates into @brainst0rm/archetype-{msp,saas-platform}
Plan item 4. Templates now ship as separate packages so they can version
independently of the CLI. Adds StarterTemplate types to @brainst0rm/config
as the shared contract; the CLI's harness-templates.ts becomes a thin
registry that imports from both archetype packages.

Verified: bare init produces 4 files (Decision #2 progressive bootstrap);
init --template saas-platform produces 16; init --template msp produces 16.
